add Demo Project to README.md
Basic demo project to show how to create a QuickParseCSV object, profile a CSV file, and read it into an arrayList of the mapped object type. Does not include maven dependancies (not yet in existence), and in the future I want to include a CSV file so that "GeneratedClass" can be more concrete. Its a start.