Add parse specifier to improve localDateTime, Boolean performance
Biggest issue right now is performance of datetimes and booleans on the "With class" load. It is too slow, >5 seconds. No suprise due to the amount of parsing required. 

This aims to fix that (and does to an extent vs 2-18 benchmark) by storing the parse format of the last seen parse for those types, while defaulting to the normal code if that fails, protecting guarantee of correct parse.